refactor: CircleCI のビルドエラー対応のため、一旦 namespace を英語に変更
