+++ Reservation/Domain/Reservations/ReserverId.cs
@@ -1,4 +1,4 @@
-namespace Reservation.Domain.予約 {
+namespace Reservation.Domain.Reservations {
     // 予約者ID
     public class ReserverId {
