+++ Reservation/Infrastructure/予約希望Repository.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
-using Reservation.Domain.予約;
-using Reservation.Domain.予約.会議室;
-using Reservation.Domain.予約.期間;
+using Reservation.Domain.Reservations;
+using Reservation.Domain.Reservations.MeetingRooms;
+using Reservation.Domain.Reservations.Period;
 
 namespace Reservation.Infrastructure
 {
