+++ Test/予約希望RepositoryTest.cs
@@ -1,9 +1,9 @@
 using System;
 using Xunit;
 using Reservation.Domain;
-using Reservation.Domain.予約;
-using Reservation.Domain.予約.会議室;
-using Reservation.Domain.予約.期間;
+using Reservation.Domain.Reservations;
+using Reservation.Domain.Reservations.MeetingRooms;
+using Reservation.Domain.Reservations.Period;
 using Reservation.Infrastructure;
 
 namespace Test
